FEAT: Adicionado tela de loading no get

+++ frontend/src/pages/Home/FormSuccess/FormSuccess.tsx
@@ -4,6 +4,9 @@ import getFirstName from "../../../utils/getFirstName";
 
 import {
   ContentContainer,
+  FetchingContainer,
+  FetchingImg,
+  FetchingTitle,
   FigureContainer,
   FormSuccessButton,
   FormSuccessContainer,
@@ -13,11 +16,15 @@ import {
   InfoDescription,
   InfoTitle,
   InfosContainer,
+  Progress,
 } from "./styles";
 
 import success_img from "../../../assets/success_image.png";
 import { useEffect, useState } from "react";
 import Modal from "../../../components/Modal/Modal";
+
+import loading_img from "../../../assets/loading_image.jpg";
+import { CircularProgress } from "@mui/material";
 
 type FormSuccessProps = {
   handleOnClick: () => void;
@@ -34,20 +41,28 @@ function FormSucess({ handleOnClick }: FormSuccessProps) {
     async function fetchDriverRegister() {
       async function fetchData() {
         setFetching(true);
-        const res = await fetch("http://localhost:3000/driver/185");
+        const res = await fetch("http://localhost:3000/driver/1");
         const data = await res.json();
         return data;
       }
 
       try {
-        const driverRegister = await fetchData();
-        setData(driverRegister);
+        const data = await fetchData();
+        setData(data);
       } catch (error) {
         setError(true);
       }
 
-      setFetching(false);
+      setTimeout(() => {
+        setFetching(false);
+      }, 2000);
+    }
 
+    fetchDriverRegister();
+  }, []);
+
+  useEffect(() => {
+    if (data) {
       try {
         fetch("http://localhost:3000/driver/1", {
           method: "DELETE",
@@ -56,13 +71,16 @@ function FormSucess({ handleOnClick }: FormSuccessProps) {
         console.log(error);
       }
     }
-
-    fetchDriverRegister();
-  }, []);
+  }, [data]);
 
   return (
     <FormSuccessContainer>
-      {fetching && <p>Is Fetching</p>}
+      {fetching && (
+        <FetchingContainer>
+          <FetchingTitle>Fetching driver dates</FetchingTitle>
+          <Progress />
+        </FetchingContainer>
+      )}
 
       {!fetching && error && (
         <Modal fn={handleOnClick} initialOpen={error}>
@@ -74,9 +92,7 @@ function FormSucess({ handleOnClick }: FormSuccessProps) {
         <>
           <FormSuccessTitleWrapper>
             <Check />
-            <FormSuccessTitle>
-              Welcome, {getFirstName(data?.fullName)}
-            </FormSuccessTitle>
+            <FormSuccessTitle>Welcome, Felipe</FormSuccessTitle>
           </FormSuccessTitleWrapper>
           <ContentContainer>
             <InfosContainer>
